Revert "Added bank transfer"
