+++ locales/en/common.json
@@ -54,29 +54,5 @@
   "purchase-fail-message": "Your tickets purchase is not finished yet. Please start over from beginning.",
   "go-home": "Go to home",
   "welcome": "Welcome",
-  "no-tickets": "You do not have any tickets available.",
-  "show-payment-details": "Show payment details",
-  "hide-payment-details": "Hide payment details",
-  "pd-amount-received": "Amount Received",
-  "pd-transfer-amount": "Transfer Amount",
-  "pd-show-full-transfer-details": "Show full transfer details",
-  "pd-hide-full-transfer-details": "Hide full transfer details",
-  "pd-account_holder_name": "Account holder name",
-  "pd-account_number": "Account number",
-  "pd-account_type": "Account type",
-  "pd-bank_code": "Bank code",
-  "pd-bank_name": "Bank name",
-  "pd-branch_code": "Branch code",
-  "pd-branch_name": "Branch name",
-  "pd-message": "Please transfer {{amount}} to the above bank account to complete your payment.",
-  "pd-show": "Show payment details",
-  "pd-hide": "Hide payment details",
-  "ordered-at": "Ordered At",
-  "ts-authorized": "Awaiting payment",
-  "ts-captured": "Completed",
-  "ts-expired": "Expired",
-  "ts-failed": "Failed",
-  "ts-cancelled": "Cancelled",
-  "ts-created": "Pending",
-  "awaiting-payment": "Awaiting payment"
+  "no-tickets": "You do not have any tickets available."
 }
