+++ locales/jp/common.json
@@ -54,27 +54,5 @@
   "purchase-fail-message": "チケットの購入はまだ完了していません。 最初からやり直してください。",
   "go-home": "トップページに戻る",
   "welcome": "ようこそ",
-  "no-tickets": "利用可能なチケットがありません。",
-  "pd-amount-received": "受け取った金額",
-  "pd-transfer-amount": "払込金額",
-  "pd-show-full-transfer-details": "振込の詳細を表示",
-  "pd-hide-full-transfer-details": "振込の詳細を非表示",
-  "pd-account_holder_name": "口座名義",
-  "pd-account_number": "口座番号",
-  "pd-account_type": "口座の種類",
-  "pd-bank_code": "金融機関コード",
-  "pd-bank_name": "銀行名",
-  "pd-branch_code": "支店コード",
-  "pd-branch_name": "支店名",
-  "pd-message": "上記銀行口座に{{amount}}をお振込みいただきお支払いを完了してください。",
-  "pd-show": "支払詳細を表示",
-  "pd-hide": "支払詳細を非表示",
-  "ordered-at": "注文日時",
-  "ts-authorized": "支払い待ち",
-  "ts-captured": "支払い完了",
-  "ts-expired": "期限が切れました",
-  "ts-failed": "失敗しました",
-  "ts-cancelled": "キャンセル",
-  "ts-created": "保留中",
-  "awaiting-payment": "支払い待ち"
+  "no-tickets": "利用可能なチケットがありません。"
 }
