+++ package.json
@@ -8,10 +8,7 @@
     "start": "next start",
     "lint": "next lint",
     "preview": "next build && next start",
-    "schema:migrate": "npx prisma migrate dev",
-    "schema:deploy": "npx prisma migrate deploy",
-    "stripe:webhook": "/usr/local/bin/stripe listen --forward-to http://localhost:3000/api/webhook/stripe --forward-connect-to http://localhost:3000/api/webhook/stripe",
-    "postinstall": "npx prisma generate"
+    "stripe:webhook": "/usr/local/bin/stripe listen --forward-to http://localhost:3000/api/webhook/stripe --forward-connect-to http://localhost:3000/api/webhook/stripe"
   },
   "dependencies": {
     "@aws-sdk/client-sns": "^3.298.0",
