+++ pages/user/my-tickets.tsx
@@ -1,16 +1,19 @@
 import Head from "next/head"
 import { Layout, Ticket, UserLayout } from "@/src/components"
-import { listAvailableTickets, listPendingTransactions } from "@/src/utils/user"
+import { DB } from "@/src/utils/db"
+import { eventById, ticketById } from "@/src/utils/temp"
+import { KomojuStatus } from "@prisma/client"
 import _ from "lodash"
+import moment from "moment"
 import { getServerSession } from "next-auth/next"
+import { useSession } from "next-auth/react"
 import useTranslation from "next-translate/useTranslation"
 
 import { siteConfig } from "@/src/config/site"
-import { AwaitingPayment } from "@/src/components/AwaitingPayment"
 import { Button } from "@/src/components/ui/button"
 import { authOptions } from "../api/auth/[...nextauth]"
 
-export default function MyTicketsPage({ tickets, pending }) {
+export default function MyTicketsPage({ transactions }) {
   const { t } = useTranslation("common")
   return (
     <Layout>
@@ -19,42 +22,23 @@ export default function MyTicketsPage({ tickets, pending }) {
       </Head>
       <UserLayout>
         <div className="relative mx-auto max-w-4xl">
-          <h1 className="flex items-center justify-between text-xl font-medium tracking-tight text-slate-900">
+          <h1 className="text-xl font-medium tracking-tight text-slate-900">
             {t("ticket-your-tickets")}
-            <Button variant="subtle">{t("go-home")}</Button>
           </h1>
           <div className="mt-4 mb-6">
-            {tickets.length > 0 ? (
+            {transactions.length > 0 ? (
               <div className="space-y-4">
-                {tickets.map((transaction) => {
+                {transactions.map((transaction) => {
                   return (
                     <Ticket key={transaction.id} transaction={transaction} />
                   )
                 })}
               </div>
             ) : (
-              <p>{t("no-tickets")}</p>
-            )}
-          </div>
-          <h1 className="flex items-center justify-between text-xl font-medium tracking-tight text-slate-900">
-            {t("awaiting-payment")}
-          </h1>
-          <div className="mt-4 mb-6">
-            {pending.filter(
-              (transaction) => transaction.status === "authorized"
-            ).length > 0 ? (
-              <div className="space-y-4">
-                {pending.map((transaction) => {
-                  return (
-                    <AwaitingPayment
-                      key={transaction.id}
-                      transaction={transaction}
-                    />
-                  )
-                })}
+              <div className="space-y-5">
+                <p>{t("no-tickets")}</p>
+                <Button variant="subtle">{t("go-home")}</Button>
               </div>
-            ) : (
-              <p>{t("no-tickets")}</p>
             )}
           </div>
         </div>
@@ -66,11 +50,53 @@ export default function MyTicketsPage({ tickets, pending }) {
 export async function getServerSideProps(context) {
   const session = await getServerSession(context.req, context.res, authOptions)
   const userId = session.user.id
+  const transactions = await DB.transaction.findMany({
+    where: { userId, status: KomojuStatus.captured },
+    select: {
+      id: true,
+      eventId: true,
+      ticketTypeId: true,
+      unitPrice: true,
+      quantity: true,
+      totalPrice: true,
+      status: true,
+      createdAt: true,
+      updatedAt: true,
+      tickets: {
+        select: {
+          id: true,
+          status: true,
+          ticketTypeId: true,
+          createdAt: true,
+          upatedAt: true,
+        },
+      },
+    },
+    orderBy: {
+      createdAt: "desc",
+    },
+  })
 
-  const tickets = await listAvailableTickets(userId)
-  const pending = await listPendingTransactions(userId)
+  const populatedTransactions = transactions.map((transaction) => {
+    return {
+      ...transaction,
+      event: eventById(transaction.eventId),
+      tickets: transaction.tickets.map((ticket) => {
+        return {
+          ...ticket,
+          ..._.omit(ticketById(transaction.eventId, ticket.ticketTypeId), [
+            "id",
+          ]),
+          createdAt: moment(ticket.createdAt).unix(),
+          upatedAt: moment(ticket.upatedAt).unix(),
+        }
+      }),
+      createdAt: moment(transaction.createdAt).unix(),
+      updatedAt: moment(transaction.updatedAt).unix(),
+    }
+  })
 
   return {
-    props: { tickets, pending, session }, // will be passed to the page component as props
+    props: { transactions: populatedTransactions, session }, // will be passed to the page component as props
   }
 }
