--- src/lib/stripe/fetchNextAction.ts
@@ -1,20 +1,0 @@
-import {
-  retrieveCheckoutSession,
-  retrievePaymentIntent,
-} from "./checkoutSession"
-
-export const fetchNextAction = async (sessionId) => {
-  try {
-    if (!sessionId) {
-      return null
-    }
-    const checkoutSession = await retrieveCheckoutSession(sessionId)
-    const paymentIntent = await retrievePaymentIntent(
-      checkoutSession.payment_intent
-    )
-    return paymentIntent.next_action
-  } catch (error) {
-    console.log(error)
-    return null
-  }
-}
