--- src/utils/user/index.ts
@@ -1,2 +1,0 @@
-export * from "./listTickets"
-export * from "./listPendingTransactions"
